Major overhaul of the letsencrypt module to support ACME v2
ACME (the Automated Certificate Management Environment, [1]) is the
protocol used for certificate mamagement on letsencrypt.org. The API
version v was released on 2016 but was updated in 2018 by ACME v2, is
not backwards compatible with v1. Letsencrypt announced in march 2018
to drop the support of ACME v1 in several steps:

- Nov 2019: End of account registrations via ACME v1
- Jun 2020: End of new domain registrations via ACME v1
- Jun 2021: EOL ACME v1 certificate issuing

The new version is based on its crypto requirements solely on OpenSSL,
it uses the NaviServer builtins and as well the "openssl" binary (the
usage of tcllib pki was dropped).  To avoid potential troubles, use
this with a recent version of NaviServer (currently the tip version
form BitBucket) or with NaviServer 4.99.19 when this is released.

-g

Unfortunately, i was hit by account registrations issue
two days ago, so i had to implement v2 to be able to get new
certificates.

[1] https://en.wikipedia.org/wiki/Automated_Certificate_Management_Environment
[2] https://community.letsencrypt.org/t/end-of-life-plan-for-acmev1/88430